Remove native executor
